fix(risk): single source of truth for risk-tier cutoffs (B2)
Risk-tier score cutoffs were duplicated in three places and had already drifted:
- engine.js tierFor(): Critical>=90 / High>=70   (canonical — produces the stored
  RiskScores.riskTier and the UI badge)
- routes/riskScores.js computeTier(): Critical>=80 / High>=60   ← DIVERGED
- RiskScoreSection.jsx ScoreBar: score cutoffs for the bar colour

So an analyst override re-tiered an entity on the 80/60 scale while its batch tier
used 90/70 — an 85 was "High" from the scorer but "Critical" after an override.

Extract the cutoffs into a dependency-free app/api/src/riskscoring/tiers.js. engine.js
imports + re-exports it (internal use + backward-compatible tierFor export);
routes/riskScores.js imports it and its computeTier is deleted (both override call
sites now use tierFor) — fixing the drift without coupling the simple SELECT route
to engine.js's re2/db imports. The UI ScoreBar now colours the bar from the
backend-provided tier via the shared TIER_STYLES map (single UI tier→colour source),
so it can't drift from the badge or the engine thresholds either.

Tests: riskScores.coverage override-clear case for score 85 now asserts High (proof
the drift is fixed); engine.test tierFor unchanged (90/70); new RiskScoreSection
mount test asserts the bar colour tracks the tier. API 60/60 + UI 6/6 green; both
complexity ratchets pass.
