+++ changes/risk-tier-single-source.md
@@ -1,0 +1,2 @@
+- Fixed inconsistent risk-tier labels after an analyst override. The override path used a different Critical/High threshold (80/60) than the scoring engine (90/70), so overriding a score could re-tier an entity on a different scale than the batch run that produced its stored tier. Both now share a single source of truth, so a score of, say, 85 is labelled "High" everywhere. (Entities with scores between 60–89 may now show a corrected tier.)
+- The risk-score bar colour now follows the entity's risk tier, matching the tier badge, instead of a separate hardcoded set of score cutoffs.
